requirements.txt
python-telegram-bot==20.8
anthropic>=0.40.0
requests
beautifulsoup4
python-dotenv
aiohttp
